backend/requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary
Pillow